PartID issues and chart updates

+++ backend/jum-api/jumwebapi/Features/Users/Queries/GetUserByPartId.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace jumwebapi.Features.Users.Queries;
 
-public sealed record GetUserByPartId(long partId) : IRequest<JustinUser>;
+public sealed record GetUserByPartId(decimal partId) : IRequest<JustinUser>;
 public class GetUserByPartIdHandler : IRequestHandler<GetUserByPartId, JustinUser>
 {
     private readonly IUserService _userService;
